requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
lxml==4.5.0
Markdown==3.2.2
multidict==4.7.5
oauth2client==4.1.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
six==1.14.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
zipp==3.1.0
